refactored the injections into DB
