+++ new_arrivals_chi/app/authorize_routes.py
@@ -27,8 +27,9 @@ Creation:
 from flask import Blueprint, render_template, redirect, url_for, request, flash
 from werkzeug.security import generate_password_hash, check_password_hash
 from new_arrivals_chi.app.database import db, User
-from new_arrivals_chi.app.utils import validate_email_syntax, validate_password, extract_signup_data, create_user
+from new_arrivals_chi.app.utils import validate_email_syntax, validate_password, extract_signup_data, extract_new_pw_data
 from flask_login import login_user, login_required, logout_user, current_user
+from data_handler import create_user, change_db_password
 
 authorize = Blueprint("authorize", __name__, static_folder="/static")
 
@@ -74,7 +75,7 @@ def signup_post():
         flash("Passwords do not match. Try again")
 
     elif not validate_password(password):
-        flash("Please enter a valid password")
+        flash("Your Password does not meet our strength requirements.")
        
     else:
         # Meets all sign up requirements
@@ -162,25 +163,25 @@ def post_change_password():
         otherwise redirects back to the change password page with a flash
         message.
     """
-    old_password = request.form.get("old_password")
-    new_password = request.form.get("new_password")
-    new_password_confirm = request.form.get("new_password_confirm")
+
+    old_password, new_password, new_password_confirm = extract_new_pw_data(request.form)
+
 
     if not check_password_hash(current_user.password, old_password):
         flash("Wrong existing password. Try again")
-    elif (
-        old_password == new_password
-    ):  # Do not need to check password hash because old password is correct
+
+    elif (old_password == new_password ):  
+        # Do not need to check password hash because old password is correct
         flash("New password cannot be the same as your previous password.")
+
     elif not new_password == new_password_confirm:
         flash("New passwords do not match. Try again")
+
     elif not validate_password(new_password):
-        flash("New password does not meet requirements. Try again")
+        flash("Your Password does not meet our strength requirements. Try again")
+
     else:
-        current_user.password = generate_password_hash(
-            new_password, method="pbkdf2:sha256"
-        )
-        db.session.commit()
+        change_db_password(new_password)
         flash("Password change successful.")
         return redirect(url_for("main.profile"))
